Add EDA solutions for five programming tasks
Added implementations for the following problems: Count Number of Teams, Counting Bits, Maximum Product of Two Elements in an Array, Number of Steps to Reduce a Number to Zero, and Number of Students Doing Homework at a Given Time. Each file contains a Python class with a method solving the respective problem.